requirements.txt
streamlit
gspread
google-auth
requests
